application/requirements.txt
Flask==3.1.0
Flask-MySQLdb==2.0.0
gunicorn==23.0.0
markupsafe==2.1.5
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
itsdangerous>=2.2
Flask-Mail==0.9.1
google-genai==1.14.0
python-dotenv==1.1.0
